got most of the license code working

+++ LicenseManager/src/com/intrinsarc/licensemanager/LicenseViewer.java
@@ -4,6 +4,8 @@ import java.awt.*;
 import java.awt.event.*;
 import java.io.*;
 import java.security.*;
+import java.text.*;
+import java.util.*;
 
 import javax.swing.*;
 
@@ -17,13 +19,22 @@ public class LicenseViewer
 			final PublicKey publicKey,
 			LDetails license,
 			final File current,
+			boolean isNew,
 			final Runnable refresher)
 	{
 		JPanel panel = new JPanel(new BorderLayout());
 		panel.add(new JLabel("<html><b>License editor"), BorderLayout.NORTH);
 
 		final DetailsViewer editor = new DetailsViewer(false);
-		JComponent main = editor.makeViewer(license);
+		final JButton save = new JButton("Save");
+		save.setEnabled(isNew);
+		JComponent main = editor.makeViewer(license, new Runnable()
+		{
+			public void run()
+			{
+				save.setEnabled(true);
+			}
+		});
 		panel.add(main, BorderLayout.CENTER);
 
 		JPanel buttons = new JPanel();
@@ -32,7 +43,6 @@ public class LicenseViewer
 		JPanel group = new JPanel();
 		buttons.add(group, BorderLayout.WEST);
 
-		JButton save = new JButton("Save");
 		buttons.add(save);
 		save.addActionListener(new ActionListener()
 		{
@@ -44,12 +54,19 @@ public class LicenseViewer
 				{
 					String name = get(editor, "email") + "-" + get(editor, "number") + ".license";
 					File file = new File(current, name);
+					if (file.exists())
+					{
+						int option = JOptionPane.showConfirmDialog(parent, "License exists already. Overwrite?");
+						if (option != 0)
+							return;
+					}
 					FileWriter writer = new FileWriter(file);
 					BufferedWriter buf = new BufferedWriter(writer);				
 					LDetails real = editor.makeDetails();
 					buf.write(real.toString());
 					buf.close();
-					writer.close();					
+					writer.close();	
+					save.setEnabled(false);
 				}
 				catch (Exception ex)
 				{
@@ -72,34 +89,56 @@ public class LicenseViewer
 				// private key
 				// and then turn it into a details structure and encrypt using the
 				// private key
-				String encmacs = editor.getField("encmacs").getText().trim();
-				JTextField decmacs = editor.getField("decmacs");
-				if (encmacs.length() > 0)
+				String machineId = editor.getField("machine-id").getText().trim();
+				JTextField macs = editor.getField("macs");
+				if (machineId.length() > 0)
 				{
 					try
 					{
-						// decrypt
-						editor.getField("decmacs").setText(LWork.decrypt(encmacs, null, privateKey));
-					} catch (Exception ex)
+						String hex = "";
+						String actual = Finder.toHex(Utils.decode(machineId));
+						for (int lp = 0; lp < actual.length() / 12; lp++)
+						{
+							if (lp != 0)
+								hex += ",";
+							hex += actual.substring(lp * 12, lp * 12 + 12);
+						}
+						macs.setText(hex);
+					}
+					catch (Exception ex)
 					{
-						JOptionPane.showMessageDialog(parent, ex.getMessage(), "Cannot decoded encrypted macs",
+						JOptionPane.showMessageDialog(parent, ex.getMessage(), "Cannot decoded base64 macs",
 								JOptionPane.ERROR_MESSAGE);
 						return;
 					}
 				}
 
-				// we must have some decmacs
-				String macval = decmacs.getText().trim();
+				// we must have some macs
+				String macval = macs.getText().trim();
 				if (macval.length() == 0)
 				{
-					JOptionPane.showMessageDialog(parent, "No decrypted mac addresses!", "Licensing problem",
+					JOptionPane.showMessageDialog(parent, "No mac addresses!", "Licensing problem",
 							JOptionPane.ERROR_MESSAGE);
 					return;
 				}
 
 				// make a new string and encrypt it
-				String lic = "user=" + get(editor, "user") + "\n" + "email=" + get(editor, "email") + "\n" + "number="
-						+ get(editor, "number") + "\n" + "macs=" + macval;
+				int expiryDays = new Integer(get(editor, "expiry-days"));
+				String expiry = "never";
+				if (expiryDays >= 0)
+				{
+					Calendar cal = Calendar.getInstance();
+					cal.add(Calendar.DATE, expiryDays);
+					DateFormat formatter = DateFormat.getDateInstance(DateFormat.DEFAULT, new Locale("en", "US"));
+					expiry = formatter.format(cal.getTime());
+				}
+
+				String lic =
+					"user=" + get(editor, "user") +
+					"\nemail=" + get(editor, "email") +
+					"\nnumber=" + get(editor, "number") +
+					"\nexpiry=" + expiry +
+					"\nmachine-id=" + get(editor, "machine-id");
 
 				// encrypt it and get the bytes
 				try
@@ -128,34 +167,16 @@ public class LicenseViewer
 
 					JPanel panel = new JPanel(new BorderLayout());
 					DetailsViewer viewer = new DetailsViewer(true);
-					panel.add(viewer.makeViewer(new LDetails(lic)), BorderLayout.NORTH);
-					JOptionPane.showConfirmDialog(parent, panel, "License details", JOptionPane.DEFAULT_OPTION,
+					panel.add(viewer.makeViewer(new LDetails(lic), null), BorderLayout.NORTH);
+					JOptionPane.showConfirmDialog(
+							parent,
+							panel,
+							"License details",
+							JOptionPane.DEFAULT_OPTION,
 							JOptionPane.INFORMATION_MESSAGE);
 				} catch (Exception ex)
 				{
 					JOptionPane.showMessageDialog(parent, ex.getMessage(), "Cannot decrypt license", JOptionPane.ERROR_MESSAGE);
-				}
-			}
-		});
-
-		JButton macs = new JButton("Local macs");
-		buttons.add(macs);
-		macs.addActionListener(new ActionListener()
-		{
-			@Override
-			public void actionPerformed(ActionEvent arg0)
-			{
-				String macs = "";
-				for (String str : Finder.findAll())
-					if (macs.length() == 0)
-						macs += str;
-					else macs += "," + str;
-				try
-				{
-					editor.getField("encmacs").setText(LWork.encrypt(macs, publicKey, null));
-				} catch (Exception ex)
-				{
-					JOptionPane.showMessageDialog(parent, ex.getMessage(), "Cannot encrypt macs", JOptionPane.ERROR_MESSAGE);
 				}
 			}
 		});
